Move `neurom.fst._dendrogram` into `neurom.view` module.
