+++ neurom/view/_compat.py
@@ -28,7 +28,7 @@
 
 '''Old-style new-style neurite compatibility hacks'''
 
-from neurom.fst._dendrogram import Dendrogram
+from neurom.view._dendrogram import Dendrogram
 from neurom import iter_neurites
 from neurom.point_neurite import segments as seg
 from neurom.point_neurite.point_tree import PointTree
