+++ neurom/view/_dendrogram.py
